// src/interfaces.ts
export interface CloudTableProps {
  width: number;
  height: number;
  rows: Array<any>;
  rowCount: number;
  columns: Array<IColumn>;
  onLoadMore: Function;
  onColumnResize: Function
}

export interface IColumn {
  // Header label
  label: string;
  // Column value
  key: string;
  // Column width
  width: number;
  // Read only field
  readOnly: Boolean;
  type: string;
  view: React.ComponentClass<any>,
  viewOption?: Object,
  inlineEditor: React.ComponentClass<any>,
  expandEditor: React.ComponentClass<any>,
  onChange: Function
}